split tests for components
