+++ performance/src/test/com/clickhouse/benchmark/BenchmarkRunner.java
@@ -1,9 +1,11 @@
 package com.clickhouse.benchmark;
 
 
-import com.clickhouse.benchmark.clients.Components;
+import com.clickhouse.benchmark.clients.Compression;
+import com.clickhouse.benchmark.clients.Deserializers;
 import com.clickhouse.benchmark.clients.InsertClient;
 import com.clickhouse.benchmark.clients.QueryClient;
+import com.clickhouse.benchmark.clients.Serializers;
 import org.openjdk.jmh.annotations.Mode;
 import org.openjdk.jmh.profile.GCProfiler;
 import org.openjdk.jmh.profile.MemPoolProfiler;
@@ -32,7 +34,9 @@ public class BenchmarkRunner {
         Options opt = new OptionsBuilder()
                 .include(QueryClient.class.getSimpleName())
                 .include(InsertClient.class.getSimpleName())
-//                .include(Components.class.getSimpleName())
+                .include(Compression.class.getSimpleName())
+                .include(Serializers.class.getSimpleName())
+                .include(Deserializers.class.getSimpleName())
                 .forks(1) // must be a fork. No fork only for debugging
                 .mode(Mode.SampleTime)
                 .timeUnit(TimeUnit.MILLISECONDS)
